Rename function

+++ innsender/src/main/kotlin/no/nav/soknad/innsending/rest/ekstern/EksternRestApi.kt
@@ -78,7 +78,7 @@ class EksternRestApi(
 		)
 
 		val ettersending = soknadService.hentSoknad(innsendingsId)
-		tilgangskontroll.validerSoknadsTilgang(ettersending)
+		tilgangskontroll.validateSoknadAccess(ettersending)
 
 		soknadService.deleteSoknadFromExternalApplication(ettersending, applikasjon)
